pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudways-api"
version = "0.1.0"
description = "Docker-containerized CLI for Cloudways API operations"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
cloudways = "cloudways_api.cli:app"

[tool.setuptools.packages.find]
include = ["cloudways_api*"]

[tool.setuptools.package-data]
"cloudways_api" = ["templates/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]
testpaths = ["tests"]

[tool.ruff]
exclude = [".claude/hooks", "scripts/prism"]

[tool.coverage.run]
source = ["cloudways_api"]
omit = [
    "cloudways_api/__main__.py",
]

[tool.coverage.report]
fail_under = 90
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
